postgers replaced sqlite

+++ database/migrations/20211009115632_live_trade/postgres.sql
@@ -1,11 +1,11 @@
 -- +goose Up
 CREATE TABLE live_trade (
-    id   SERIAL PRIMARY KEY,
+    id SERIAL PRIMARY KEY,
     side text not null,
     entry_order_id text NOT NULL,
-    entry_price REAL NOT NULL,
+    entry_price double precision NOT NULL,
     entry_time TIMESTAMPTZ NOT NULL,
-    stop_loss_price REAL NOT NULL,
+    stop_loss_price double precision NOT NULL,
     strategy_id TEXT NOT NULL,
     status TEXT NOT NULL,
     pair text NOT NULL,
@@ -13,11 +13,10 @@ CREATE TABLE live_trade (
     exit_time TIMESTAMPTZ ,
     take_profit_price REAL,
     profit_loss_points REAL,
-    exit_price REAL,
+    exit_price double precision,
 
     created_at       TIMESTAMP NOT NULL DEFAULT (now() at time zone 'utc'),
     updated_at       TIMESTAMP NOT NULL DEFAULT (now() at time zone 'utc')
 );
 -- +goose Down
 DROP TABLE live_trade;
-
